+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Set user-agent to `giantswarm-capa-operator` when making requests to the k8s API.
+
 ## [0.11.0] - 2023-10-04
 
 ### Changed